requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-ckeditor>=6.7.0
